style: clean and document the code
