--- std/algebra/emulated/fields_bw6761/frobenius.go
@@ -1,27 +1,0 @@
-package fields_bw6761
-
-import (
-	"github.com/consensys/gnark/std/math/emulated"
-)
-
-var (
-	_frobA  = emulated.ValueOf[emulated.BW6761Fp]("4922464560225523242118178942575080391082002530232324381063048548642823052024664478336818169867474395270858391911405337707247735739826664939444490469542109391530482826728203582549674992333383150446779312029624171857054392282775648")
-	_frobB  = emulated.ValueOf[emulated.BW6761Fp]("1968985824090209297278610739700577151397666382303825728450741611566800370218827257750865013421937292370006175842381275743914023380727582819905021229583192207421122272650305267822868639090213645505120388400344940985710520836292650")
-	_frobC  = emulated.ValueOf[emulated.BW6761Fp]("4922464560225523242118178942575080391082002530232324381063048548642823052024664478336818169867474395270858391911405337707247735739826664939444490469542109391530482826728203582549674992333383150446779312029624171857054392282775649")
-	_frobAC = emulated.ValueOf[emulated.BW6761Fp]("-1")
-	_frobBC = emulated.ValueOf[emulated.BW6761Fp]("1968985824090209297278610739700577151397666382303825728450741611566800370218827257750865013421937292370006175842381275743914023380727582819905021229583192207421122272650305267822868639090213645505120388400344940985710520836292651")
-)
-
-// Frobenius set z in E6 to Frobenius(x), return z
-func (e Ext6) Frobenius(x *E6) *E6 {
-	var z E6
-	z.B0.A0 = x.B0.A0
-	z.B0.A1 = *e.fp.Mul(&x.B0.A1, &_frobA)
-	z.B0.A2 = *e.fp.Mul(&x.B0.A2, &_frobB)
-
-	z.B1.A0 = *e.fp.Mul(&x.B1.A0, &_frobC)
-	z.B1.A1 = *e.fp.Mul(&x.B1.A1, &_frobAC)
-	z.B1.A2 = *e.fp.Mul(&x.B1.A2, &_frobBC)
-
-	return &z
-}
